main.cpp and run.cpp
