+++ atomtex_spe_file_cli/CMakeLists.txt
@@ -1,6 +1,6 @@
 find_package(Boost REQUIRED COMPONENTS program_options json)
 
-add_executable(atomtex_spe_file_cli main.cpp)
+add_executable(atomtex_spe_file_cli main.cpp run.cpp)
 target_link_libraries(atomtex_spe_file_cli PRIVATE
   Boost::program_options
   Boost::json
