+++ atomtex_spe_file_cli/run.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace atomtex_spe_file {
+namespace cli {
+
+int run(const int argc, const char* argv[]);
+    
+}    
+}
